refactor(PurchasePrice): 클래스명 변경 & 불필요 클래스 정리
